Refactor : LottoGenerator 생성
