+++ src/main/java/lotto/model/LottoPurchaser.java
@@ -2,20 +2,21 @@ package lotto.model;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.model.util.LottoGenerator;
 
 public class LottoPurchaser {
 
-    private final AutoLottoGenerator autoLottoGenerator;
+    private final LottoGenerator lottoGenerator;
 
-    public LottoPurchaser(final AutoLottoGenerator autoLottoGenerator) {
-        this.autoLottoGenerator = autoLottoGenerator;
+    public LottoPurchaser(final LottoGenerator autoLottoGenerator) {
+        this.lottoGenerator = autoLottoGenerator;
     }
 
     public PurChasedLotto purchaseLotto(final LottoMoney lottoMoney) {
         final int lottoCount = lottoMoney.getLottoCount();
         final List<Lotto> lottos = new ArrayList<>();
         for (int i = 0; i < lottoCount; i++) {
-            lottos.add(autoLottoGenerator.generate());
+            lottos.add(new Lotto(lottoGenerator.generate()));
         }
         return new PurChasedLotto(lottos);
     }
