+++ src/main/java/lotto/model/util/LottoGenerator.java
@@ -1,0 +1,10 @@
+package lotto.model.util;
+
+import java.util.List;
+import lotto.model.Lotto;
+import lotto.model.LottoNumber;
+
+public interface LottoGenerator {
+
+    List<LottoNumber> generate();
+}
