+++ src/test/java/lotto/model/AutoLottoGeneratorTest.java
@@ -1,5 +1,7 @@
 package lotto.model;
 
+import lotto.model.util.AutoLottoGenerator;
+import lotto.model.util.LottoGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,10 +13,10 @@ public class AutoLottoGeneratorTest {
     @DisplayName("로또 자동 생성기는 올바른 로또를 생성한다.")
     void 자동_로또_생성() {
 
-        AutoLottoGenerator autoLottoGenerator = new AutoLottoGenerator();
+        LottoGenerator autoLottoGenerator = new AutoLottoGenerator();
 
         assertThatCode(autoLottoGenerator::generate)
-                .doesNotThrowAnyException();
+            .doesNotThrowAnyException();
     }
 
 }
